+++ Extension/pages/content-script-start/mv2.ts
@@ -21,5 +21,9 @@ import '@adguard/tswebextension/content-script';
 import { ContentUtils } from '../../src/content-script/content-utils/main';
 import { SubscribeToScriptlets } from '../../src/content-script/subscribe-to-scriptlets';
 
+/**
+ * Following methods are async BUT called without await **intentionally*
+ * because if they are called as sync, they can slow down frames loading in Firefox.
+ */
 ContentUtils.init();
 SubscribeToScriptlets.init();