+++ tools/bundle/firefox/manifest.firefox.ts
@@ -22,13 +22,21 @@ import {
     FIREFOX_WEBEXT_UPDATE_URL,
 } from '../../constants';
 import { OPTIONS_PAGE } from '../../../Extension/src/common/constants';
-import { BACKGROUND_OUTPUT, POPUP_OUTPUT } from '../../../constants';
+import {
+    BACKGROUND_OUTPUT,
+    MIN_SUPPORTED_VERSION,
+    POPUP_OUTPUT,
+} from '../../../constants';
 
 const appId = FIREFOX_APP_IDS_MAP[BUILD_ENV];
 
 if (appId === undefined) {
     throw new Error(`App ID not found for BUILD_ENV: ${BUILD_ENV}`);
 }
+
+const MIN_SUPPORTED_DESKTOP_VERSION_STR = `${String(MIN_SUPPORTED_VERSION.FIREFOX)}.0`;
+
+const MIN_SUPPORTED_ANDROID_VERSION_STR = `${String(MIN_SUPPORTED_VERSION.FIREFOX_MOBILE)}.0`;
 
 export const firefoxManifest = {
     'background': {
@@ -49,10 +57,10 @@ export const firefoxManifest = {
     'browser_specific_settings': {
         'gecko': {
             'id': appId,
-            'strict_min_version': '78.0',
+            'strict_min_version': MIN_SUPPORTED_DESKTOP_VERSION_STR,
         },
         'gecko_android': {
-            'strict_min_version': '113.0',
+            'strict_min_version': MIN_SUPPORTED_ANDROID_VERSION_STR,
         },
     },
     'options_ui': {